requirements.txt
pandas>=1.5.0
pyyaml>=6.0
pytest>=7.0
openpyxl>=3.0
xlrd>=2.0
pyarrow>=12.0
pydantic>=2.0
scipy>=1.10.0
